Se cambio el icono de logout por un menu desplegable en el nombre de usuario en el que puede ver su perfil o cerrar sesion desde ahi, se creo la vista user_show.php, ya se ven los datos de usuario, falta darle funcionalidad para poder editarlos

+++ php/comment.php
@@ -136,8 +136,8 @@ class Comment extends Database{
    }
 
 
-   public function delete($id){
-      $query = "DELETE FROM comments WHERE `comments`.`id` = {$id}";
+   public function delete($id, $user_id){
+      $query = "DELETE FROM comments WHERE `comments`.`id` = {$id} AND `comments`.`user_id`";
       $this->connect();
       $this->executeQuery($query, 'Comentario eliminado', 'No se puede eliminar el comentario');
       $this->disconnect();
